Update information about PHPCompatibility
The repository for the PHPCompatibility project has moved to its own GitHub organisation.

Within that organisation, CMS-specific repos have been created which contain custom rulesets based on PHPCompatibility which take back-fills/poly-fills included in the CMS into account.

It is recommended for projects which expect to be run within a WordPress context - i.e. WordPress itself, plugins and themes -, to use the WP specific PHPCompatibility ruleset to reduce false positives.

For more information:
* https://github.com/PHPCompatibility/PHPCompatibilityWP
* https://github.com/PHPCompatibility/PHPCompatibility/releases/tag/8.2.0